Move prime curve signature rendering into n20_x509_cert (#69)
Move rendering of P-256 and P-384 curve signatures into n20_x509_cert.
Previously, the interface took the signature bits as-is. This was a
purist approach to the x509 interface.
However for the sake of interoperability with the CWT/COSE rendering it
is just more convenient to have the respective certificate render
function convert the signature from a common intermediate format.